refactor: move registerCustomKeysets to a separate file
